fix(terminal): terminal microservice bug fix, trow exception for fimrware that doen't exist

+++ backend/terminal/src/main/java/com/pos/terminal/terminal/TerminalRequest.java
@@ -38,7 +38,7 @@ public class TerminalRequest {
     private String macAddress;
 
     @NotNull(message = "Current Firmware Id Type is mandatory")
-    private BigInteger currentFirmwareId;
+    private Long currentFirmwareId;
 
     @NotBlank(message = "Status Type is mandatory")
     private String status;
